Fix: Ignore hostname match, connect before login
Adds an --insecure option that disables the match_hostname check – this is
useful e.g. if you run pybofh on the same computer that runs bofhd with
encryption.

Separates between connect and login. Now we run: 1) connect, 2) get motd 3)
login. This way the user doesn't have to enter their password until a connection
is made.