Fix an issue with missing logging handler.
After introducing loggers to the `bofh` modules, a warning may be
emitted if a bofh-module is used without first configuring the python
`logging` module.

Logging is configured if the `bofh.cli` entry point is started with a
`-v`, `--verbose` flag.  However, if no such flag is given, no handlers
are configured for the python logging hierarchy.  This fix adds a
default `logging.NullHandler` to the root logger if no other logging is
configured.