Improve format suggestion parsing
The previous commit refactored the format suggestion parsing to make it
more testable and readable.  This commit changes the format suggestion
parsing so that all the relevant data structures are checked when
`bofh.formatting.get_formatter()` is called.  It also fixes a bug that
was introduced in the refacored code, where parts of the response were
replaced by tracebacks from the formatter.