Fix Pybofh repl to handle range/comma separated values for specific commands
Fixed how the repl parse values for selection, it will now pass the string to cerebrum if the
are multiple values in input (e.g. ranges and comma separated values).
The range/comma separated parsing is handled in Cerebrum on a per command basis.
An example of this is in "misc print_passwords".

Issue:PYBOFH-15