setup: target Python versions 3.6, 3.8, and 3.9
The default Python version on Red Hat Enterprise Linux is 3.6 so
it would make sense to ensure this specific Python runtime is tested.

The 3.8 runtime is popular on modern testing distributions, and 3.9
is the latest stable version which we ought to support.

All tests are passing in tox on all runtime versions.

Fixes: PYBOFH-47