setup: rename package bofh
Changes the name of this package from "pybofh" to "bofh", which
incidentally matches the exposed Python module.

pybofh is not available on PyPI but bofh is, which means we are graduating
this bofh client implementation to the de facto bofh client.

This change may cause some issues with RPM packaging and publishing to
the internal USIT PyPI package index (repo.usit.uio.no).

Fixes: PYBOFH-47